@@ -1,13 +1,10 @@
 name: dgraph
-version: 1.1.1
-description: Dgraph Dart client which communicates with the server using gRPC.
-author: Marcelo Henrique Neppel <marcelo.neppel@gmail.com>
-homepage: https://github.com/marceloneppel/dgraph
+version: 1.1.2
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
-  grpc: 0.6.6
-  mutex: ^1.0.3
-  protobuf: '>=1.0.1 <2.0.0'
+  grpc: ^3.0.2
+  mutex: ^3.0.0
+  protobuf: ^2.0.0
 dev_dependencies:
   test: ^1.5.1
