@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: api.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: camel_case_types,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 // ignore_for_file: UNDEFINED_SHOWN_NAME,UNUSED_SHOWN_NAME
@@ -26,7 +26,7 @@ class Operation_DropOp extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, Operation_DropOp> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static Operation_DropOp valueOf($core.int value) => _byValue[value];
+  static Operation_DropOp? valueOf($core.int value) => _byValue[value];
 
   const Operation_DropOp._($core.int v, $core.String n) : super(v, n);
 }
@@ -48,7 +48,7 @@ class Facet_ValType extends $pb.ProtobufEnum {
 
   static final $core.Map<$core.int, Facet_ValType> _byValue =
       $pb.ProtobufEnum.initByValue(values);
-  static Facet_ValType valueOf($core.int value) => _byValue[value];
+  static Facet_ValType? valueOf($core.int value) => _byValue[value];
 
   const Facet_ValType._($core.int v, $core.String n) : super(v, n);
 }
